Ending with the management of TVET College Times (Client and Admin Side

+++ app/Http/Controllers/TvetCollegeTimesController.php
@@ -6,6 +6,8 @@ use Illuminate\Http\Request;
 use App\Models\TvetCollegeTimes;
 use File;
 use Auth;
+use Illuminate\Http\UploadedFile;
+
 class TvetCollegeTimesController extends Controller
 {
     
@@ -52,22 +54,24 @@ class TvetCollegeTimesController extends Controller
             'publicationMonth' => 'required',
             'publicationYear' => 'required',
             'title' => 'required',
-            'size' => 'required',
             'file' => 'required|file|mimes:pdf',
         ]);
 
-        $fileName = 'tvetCollegeTimes' . '-' .time().'.'.$request->file->extension();  
         
-        $request->picture->move(public_path('files'), $fileName);
         $publicationDate=$request->input('publicationMonth').' '.$request->input('publicationYear');
         $tvetcollegetimes = new TvetCollegeTimes;
-        $tvetcollegetimes->volume=$request->input('volume');
+        $tvetcollegetimes->volume="Volume ".$request->input('volume');
         $tvetcollegetimes->title=$request->input('title');
         $tvetcollegetimes->publicationDate=$publicationDate;
-        $tvetcollegetimes->title=$request->file->getSize();
-        $tvetcollegetimes->file=$fileName;
-
-        
+        if (request()->hasFile('file')){
+            $file = $request->file('file');
+            $size1=number_format((float)$file->getSize()/(1024*1024), 2, '.', '');
+            $tvetcollegetimes->size=$size1.' Mo';
+            $fileName = 'tvetCollegeTimes' . '-' .time().'.'.$file->extension();  
+            $request->file->move(public_path('files'), $fileName);
+            
+            $tvetcollegetimes->file=$fileName;
+        }
         $tvetcollegetimes->user_id=Auth::user()->id;
         $tvetcollegetimes->save();
         return redirect()->route('admintvetcollegetimes.index')->with('success','TVET College Times has been created successfully.');
@@ -90,8 +94,9 @@ class TvetCollegeTimesController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function edit(TvetCollegeTimes $tvetcollegetimes)
+    public function edit($id)
     {
+        $tvetcollegetimes = TvetCollegeTimes::find($id);
         return view('admin.tvetcollegetimes.edit',compact('tvetcollegetimes'));
     }
 
@@ -102,35 +107,37 @@ class TvetCollegeTimesController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, TvetCollegeTimes $tvetcollegetimes)
+    public function update(Request $request, $id)
     {
+        $tvetcollegetimes = TvetCollegeTimes::find($id);
+
         $request->validate([
             'volume' => 'required',
             'publicationMonth' => 'required',
             'publicationYear' => 'required',
             'title' => 'required',
-            'size' => 'required',
             'file' => 'required|file|mimes:pdf',
-        ]);
-
-        $fileName=$tvetcollegetimes->file;
-        $file_path = "files/$fileName";
-        unlink(public_path($file_path));
-
-        $fileName = 'tvetCollegeTimes' . '-' .time().'.'.$request->file->extension();  
+        ]); 
         
-        $request->picture->move(public_path('files'), $fileName);
         $publicationDate=$request->input('publicationMonth').' '.$request->input('publicationYear');
-        $tvetcollegetimes = new TvetCollegeTimes;
-        $tvetcollegetimes->volume=$request->input('volume');
+        $tvetcollegetimes->volume="Volume ".$request->input('volume');
         $tvetcollegetimes->title=$request->input('title');
         $tvetcollegetimes->publicationDate=$publicationDate;
-        $tvetcollegetimes->title=$request->file->getSize();
-        $tvetcollegetimes->file=$fileName;
+        if (request()->hasFile('file')){
+            $fileName=$tvetcollegetimes->file;
+            $file_path = "files/$fileName";
+            unlink(public_path($file_path));
 
-        
+            $file = $request->file('file');
+            $size1=number_format((float)$file->getSize()/(1024*1024), 2, '.', '');
+            $tvetcollegetimes->size=$size1.' Mo';
+            $fileName = 'tvetCollegeTimes' . '-' .time().'.'.$file->extension();  
+            $request->file->move(public_path('files'), $fileName);
+            $tvetcollegetimes->file=$fileName;
+        }
         $tvetcollegetimes->user_id=Auth::user()->id;
         $tvetcollegetimes->save();
+        
         $tvetcollegetimes->fill($request->post())->save();
         return redirect()->route('admintvetcollegetimes.index')->with('success','TVET College Times has been updated successfully.');
     }
@@ -141,9 +148,10 @@ class TvetCollegeTimesController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy(TvetCollegeTimes $tvetcollegetimes)
+    public function destroy($id)
     {
-        $fileName=$news->file;
+        $tvetcollegetimes = TvetCollegeTimes::find($id);
+        $fileName=$tvetcollegetimes->file;
         $file_path = "files/$fileName"; 
         unlink(public_path($file_path));
         $tvetcollegetimes->delete();
